+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## Unreleased
+
+* Port to the new CO APIs
+* The config option `data_service_name_active_studies` was renamed to `data_service_name_studies`
+* The command `show-active-studies` was renamed to `show-studies`
+* In the JSON output the field `activeStudies` was renamed to `studies`
+
 ## v0.1.1
 
 * Add various CLI commands for testing the API